pyproject.toml
[project]
name = "algo-trading"
version = "0.1.0"
description = ""
authors = [
    {name = "Aymane",email = "aymanecheikh2002@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "protobuf (>=3.20.3,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "ib-async (>=2.0.1,<3.0.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "lxml (>=6.0.2,<7.0.0)",
    "lightweight-charts (>=2.1,<3.0)",
    "yfinance (>=0.2.66,<0.3.0)",
    "ta-lib (>=0.6.8,<0.7.0)",
    "pywebview[qt] (>=6.1,<7.0)",
    "ib-insync (>=0.9.86,<0.10.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "scipy (>=1.16.3,<2.0.0)",
    "statsmodels (>=0.14.5,<0.15.0)",
    "kfp ==2.4.0",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "google-cloud-aiplatform (>=1.128.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "algo_trading", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
